requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pydantic[email]
python-jose